Move to spectator to collect metrics (#707)
Default binding is to NoopRegistry.

Refer to https://github.com/Netflix/spectator for more details.

+++ priam/src/main/java/com/netflix/priam/aws/S3EncryptedFileSystemMBean.java
@@ -19,9 +19,9 @@ public interface S3EncryptedFileSystemMBean {
 
     String ENCRYPTED_FILE_SYSTEM_MBEAN_NAME = "com.priam.aws.S3EncryptedFileSystemMBean:name=S3EncryptedFileSystemMBean";
 
-    int downloadCount();
+    long downloadCount();
 
-    int uploadCount();
+    long uploadCount();
 
     int getActivecount();
